Split mobile list payloads and actions out of length-budgeted files
The groups payload pushed TerminalController.swift (+66),
MobileShellComposite.swift (+49), and MobileHostService.swift (+11) over
the Swift file length budget. Move the mobile workspace-list payload
builders and the mobile group collapse handler into
TerminalController+MobileWorkspaceList.swift, the fire-and-forget
workspace mutations into MobileShellComposite+WorkspaceActions.swift,
and the advertised capability list into
MobileHostService+Capabilities.swift, then tighten the three budget
entries to the new actuals.

Also resolves group membership with a single pass over tabs instead of
one scan per group (review feedback) via mobileWorkspaceGroupPayloads.
